backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atr-workbench-backend"
version = "0.1.0"
description = "ATR workbench backend for ship detection workflows"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0",
  "fastapi>=0.115.0",
  "mlflow>=2.17.0",
  "pillow>=10.4.0",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.6.0",
  "python-multipart>=0.0.12",
  "pyyaml>=6.0.2",
  "sqlalchemy>=2.0.36",
  "ultralytics>=8.3.0",
  "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.13.0",
  "ruff>=0.8.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "ASYNC"]
ignore = ["B008"]  # FastAPI Depends/File in argument defaults is idiomatic

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
plugins = []
exclude = ["alembic/versions"]

